+++ shielder/Cargo.toml
@@ -2,8 +2,9 @@
 resolver = "2"
 
 members = [
-    "shielder/relations",
-    "shielder/contract",
+    "contract",
+    "mocked_zk",
+    "relations",
 ]
 
 [workspace.package]
@@ -17,10 +18,15 @@ version = "0.1.0"
 
 [workspace.dependencies]
 anyhow = { version = "1.0.79", default-features = false }
-drink = { version = "0.9.0" }
+drink = { version = "0.16.0" }
 halo2-base = { package = "halo2-base", git = "https://github.com/Cardinal-Cryptography/halo2-lib", branch = "aleph" }
 ink = { version = "5.0.0", default-features = false }
 poseidon = { git = "https://github.com/zemse/pse-poseidon" }
 rand = { version = "=0.8" }
 serde = { version = "=1.0", default-features = false }
 serde_json = { version = "=1.0" }
+
+
+liminal-halo2-relations = { path = "relations", version = "=0.1.0" }
+mocked_zk = { path = "mocked_zk", version = "=0.1.0", default-features = false }
+shielder-contract = { path = "contract", version = "=0.1.0" }
