--- shielder/drink_tests/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "drink_tests"
-version = "0.1.0"
-authors = ["Cardinal"]
-homepage = "https://alephzero.org"
-repository = "https://github.com/Cardinal-Cryptography/zk-apps"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-drink = { version = "0.15.0" }
-rand = { version = "0.8.5", default-features = false }
-anyhow = { version = "1.0.79", default-features = false }
-mocked_zk = { path = "../mocked_zk", default-features = false }
-shielder-contract = { path = "../contract", default-features = false, features = ["ink-as-dependency"] }
