+++ shielder/mocked_zk/Cargo.toml
@@ -1,16 +1,17 @@
 [package]
 name = "mocked_zk"
-version = "0.1.0"
-authors = ["Cardinal"]
-homepage = "https://alephzero.org"
-repository = "https://github.com/Cardinal-Cryptography/zk-apps"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+authors.workspace = true
+documentation = "https://docs.rs/?"
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+version.workspace = true
 
 [dependencies]
-ink = { version = "5.0.0", default-features = false }
-serde = { version = "1.0.197", default-features = false, features = ["derive"] }
+ink = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
 
 [features]
 default = ["std"]
